+++ karabo/test/test_skymodel.py
@@ -23,6 +23,7 @@ from karabo.data.external_data import (
     MIGHTEESurveyDownloadObject,
 )
 from karabo.simulation.sky_model import (
+    KaraboSkyModelError,
     Polarisation,
     SkyModel,
     SkyPrefixMapping,
@@ -35,6 +36,26 @@ from karabo.simulator_backend import SimulatorBackend
 @pytest.fixture(scope="function")
 def gleam() -> SkyModel:
     return SkyModel.get_GLEAM_Sky(min_freq=72e6, max_freq=80e6)
+
+
+def test_download_gleam_and_make_sky_model():
+    sky = SkyModel.get_GLEAM_Sky(min_freq=72e6, max_freq=80e6)
+    sample_prefix_mapping = SkyPrefixMapping([], [], [])
+    number_of_sky_attributes = len(sample_prefix_mapping.__dict__)
+
+    assert sky.num_sources > 0
+
+    # -1 since we do not return the source ID
+    assert sky.to_np_array().shape == (sky.num_sources, number_of_sky_attributes - 1)
+    assert sky.source_ids["dim_0"].shape[0] == sky.shape[0]  # checking source-ids
+
+
+def test_mightee():
+    _ = SkyModel.get_MIGHTEE_Sky()
+    with pytest.raises(KaraboSkyModelError):
+        _ = SkyModel.get_MIGHTEE_Sky(min_freq=3e11)
+    _ = SkyModel.get_MIGHTEE_Sky(max_freq=3e11)
+    _ = SkyModel.get_MIGHTEE_Sky(min_freq=1.3e9, max_freq=1.4e9)
 
 
 def test_filter_sky_model(gleam: SkyModel):
